[#936] Extract CurveType/ParametricCurveType and LUT16Type into tags/ (item 3)
Created `DisplayCAL/icc_profile/tags/curve.py` (`CurveType`,
`ParametricCurveType`) and `DisplayCAL/icc_profile/tags/lut.py`
(`LUT16Type`), moved out of `icc_profile/__init__.py`. `LUT16Type._apply_black()`
uses a deferred import for `_mp_apply_black`, which still lives in
`icc_profile/__init__.py` pending its own extraction into `tonemap.py` in
item 9, mirroring the pattern already used for `pool_slice`.

Also dropped imports that became genuinely unused in `__init__.py` after
this extraction (`dict_sort`, `imfile`, `Iterable`, `u8Fixed8Number`/
`_tohex`); `s15Fixed16Number`/`_tohex` were initially removed too but
restored since `worker.py` and `tests/test_ICCProfile.py` still import
them as a re-export from `icc_profile`, joining the existing deferred
F401 backlog for item 11's lint pass.
