+++ DisplayCAL/icc_profile/__init__.py
@@ -35,8 +35,6 @@ from typing import (
 )
 from weakref import WeakValueDictionary
 
-from DisplayCAL.util_dict import dict_sort
-
 if sys.platform == "win32":
     import winreg
 
@@ -80,7 +78,7 @@ except ImportError:
             return
 
     colord = Colord()
-from DisplayCAL import colormath, edid, imfile
+from DisplayCAL import colormath, edid
 from DisplayCAL.defaultpaths import ICCPROFILES, ICCPROFILES_HOME
 from DisplayCAL.options import TEST_INPUT_CURVE_CLIPPING
 from DisplayCAL.util_list import intlist
@@ -106,7 +104,7 @@ elif sys.platform == "win32":
 if TYPE_CHECKING:
     import multiprocessing
     import threading
-    from collections.abc import Iterable, Iterator
+    from collections.abc import Iterator
     from typing import BinaryIO, TextIO
 
     from DisplayCAL.worker import Worker, Xicclu  # noqa: TC004
@@ -126,8 +124,6 @@ from DisplayCAL.icc_profile.codecs import (
     s15f16_is_equal,
     s15Fixed16Number,
     s15Fixed16Number_tohex,
-    u8Fixed8Number,
-    u8Fixed8Number_tohex,
     u16Fixed16Number,
     u16Fixed16Number_tohex,
     uInt8Number,
@@ -183,6 +179,8 @@ from DisplayCAL.icc_profile.tags.base import (
     XYZNumber,
     XYZType,
 )
+from DisplayCAL.icc_profile.tags.curve import CurveType, ParametricCurveType
+from DisplayCAL.icc_profile.tags.lut import LUT16Type
 
 
 def create_RGB_A2B_XYZ(  # noqa: N802
@@ -3192,909 +3190,6 @@ class Illuminant(ADict):
         self.description = ILLUMINANTS[self.type]
 
 
-class LUT16Type(ICCProfileTag):
-    """ICC LUT16Type tag.
-
-    Args:
-        tagData (bytes): The raw tag data.
-        tagSignature (bytes): The tag signature.
-        profile (ICCProfile): The ICC profile this tag belongs to.
-    """
-
-    def __init__(
-        self,
-        tagData: None | bytes = None,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-        profile: None | ICCProfile = None,
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        self.profile = profile
-        self._matrix = None
-        self._input = None
-        self._clut = None
-        self._output = None
-        self._i = (tagData and uInt8Number(tagData[8:9])) or 0  # Input channel count
-        self._o = (tagData and uInt8Number(tagData[9:10])) or 0  # Output channel count
-        self._g = (tagData and uInt8Number(tagData[10:11])) or 0  # cLUT grid res
-        self._n = (
-            tagData and uInt16Number(tagData[48:50])
-        ) or 0  # Input channel entries count
-        self._m = (
-            tagData and uInt16Number(tagData[50:52])
-        ) or 0  # Output channel entries count
-
-    def apply_black_offset(
-        self,
-        XYZbp: tuple[float, float, float],  # noqa: N803
-        logfile: None | TextIO = None,
-        thread_abort: None | threading.Event = None,
-        abortmessage: str = "Aborted",
-    ) -> None:
-        """Apply black point offset to the cLUT.
-
-        Args:
-            XYZbp (tuple[float, float, float]): The black point offset values
-                as a tuple of three floats (X, Y, Z).
-            logfile (None | TextIO): File-like object to write progress
-                messages to. Defaults to None.
-            thread_abort (threading.Event): Event to signal thread abortion.
-                Defaults to None.
-            abortmessage (str): Message to display when the operation is
-                aborted. Defaults to "Aborted".
-        """
-        # Apply only the black point blending portion of BT.1886 mapping
-        self._apply_black(XYZbp, False, False, logfile, thread_abort, abortmessage)
-
-    def apply_bpc(
-        self,
-        bp_out: tuple[float, float, float] = (0, 0, 0),
-        weight: bool = False,
-        logfile: None | TextIO = None,
-        thread_abort: None | threading.Event = None,
-        abortmessage: str = "Aborted",
-    ) -> None:
-        """Apply black point compensation to the cLUT.
-
-        Args:
-            bp_out (tuple, optional): The black point output values as a tuple
-                of three floats (R, G, B). Defaults to (0, 0, 0).
-            weight (bool, optional): Whether to apply a weighted black point
-                compensation. Defaults to False.
-            logfile (None | TextIO): File-like object to write the log messages
-                to.
-            thread_abort (threading.Event): Event to signal thread abortion.
-            abortmessage (str): Message to display when the operation is aborted.
-        """
-        return self._apply_black(
-            bp_out, True, weight, logfile, thread_abort, abortmessage
-        )
-
-    def _apply_black(
-        self,
-        bp_out: tuple[float, float, float],
-        use_bpc: bool = False,
-        weight: bool = False,
-        logfile: None | TextIO = None,
-        thread_abort: None | threading.Event = None,
-        abortmessage: str = "Aborted",
-    ) -> None:
-        """Apply black point compensation or offset to the cLUT.
-
-        Args:
-            bp_out (tuple): The black point output values as a tuple of three
-                floats (R, G, B).
-            use_bpc (bool, optional): Whether to use black point compensation
-                (BPC) or just apply a black offset. Defaults to False.
-            weight (bool, optional): Whether to apply a weighted black point
-                compensation. Defaults to False.
-            logfile (None | TextIO): File-like object to write progress
-                messages to. Defaults to None.
-            thread_abort (None | threading.Event, optional): Event to signal
-                thread abortion. Defaults to None.
-            abortmessage (str): Message to display when the operation is
-                aborted. Defaults to "Aborted".
-
-        Raises:
-            ValueError: If the PCS is not supported or if the black point
-                output does not match the expected format.
-        """
-        pcs = self.profile and self.profile.connectionColorSpace
-        bp_row = list(self.clut[0][0])
-        wp_row = list(self.clut[-1][-1])
-        nonzero_bp = tuple(bp_out) != (0, 0, 0)
-        interp = []
-        rinterp = []
-        if not use_bpc or nonzero_bp:
-            osize = len(self.output[0])
-            omaxv = osize - 1.0
-            orange = [i / omaxv * 65535 for i in range(osize)]
-            for i in range(3):
-                interp.append(colormath.Interp(orange, self.output[i]))
-                rinterp.append(colormath.Interp(self.output[i], orange))
-            for row in (bp_row, wp_row):
-                for column, value in enumerate(row):
-                    row[column] = interp[column](value)
-        method = "apply_bpc" if use_bpc else "apply_black_offset"
-        if pcs == b"Lab":
-            bp = colormath.Lab2XYZ(*legacy_PCSLab_uInt16_to_dec(*bp_row))
-            wp = colormath.Lab2XYZ(*legacy_PCSLab_uInt16_to_dec(*wp_row))
-        elif not pcs or pcs == b"XYZ":
-            if not pcs:
-                warnings.warn(
-                    f"LUT16Type.{method}: PCS not specified, assuming XYZ",
-                    Warning,
-                    stacklevel=2,
-                )
-            bp = [v / 32768.0 for v in bp_row]
-            wp = [v / 32768.0 for v in wp_row]
-        else:
-            raise ValueError(f"LUT16Type.{method}: Unsupported PCS {pcs!r}")
-        if [round(v * 32768) for v in bp] != [round(v * 32768) for v in bp_out]:
-            D50 = colormath.get_whitepoint("D50")  # noqa: N806
-
-            from DisplayCAL.multiprocess import pool_slice
-
-            num_workers = 1 if len(self.clut[0]) < 33 else None
-
-            # if pcs != "Lab" and nonzero_bp:
-            #     bp_out_offset = bp_out
-            #     bp_out = (0, 0, 0)
-
-            if bp != bp_out:
-                self.clut = functools.reduce(
-                    operator.iadd,
-                    pool_slice(
-                        _mp_apply_black,
-                        self.clut,
-                        (
-                            pcs,
-                            bp,
-                            bp_out,
-                            wp,
-                            use_bpc,
-                            weight,
-                            D50,
-                            interp,
-                            rinterp,
-                            abortmessage,
-                        ),
-                        {},
-                        num_workers,
-                        thread_abort,
-                        logfile,
-                    ),
-                    [],
-                )
-
-        # if pcs != "Lab" and nonzero_bp:
-        # # Apply black offset to output curves
-        # out = [[], [], []]
-        # for i in range(2049):
-        # v = i / 2048.0
-        # X, Y, Z = colormath.blend_blackpoint(v, v, v, (0, 0, 0),
-        # bp_out_offset)
-        # out[0].append(X * 2048 / 4095.0 * 65535)
-        # out[1].append(Y * 2048 / 4095.0 * 65535)
-        # out[2].append(Z * 2048 / 4095.0 * 65535)
-        # for i in range(2049, 4096):
-        # v = i / 4095.0
-        # out[0].append(v * 65535)
-        # out[1].append(v * 65535)
-        # out[2].append(v * 65535)
-        # self.output = out
-
-    @property
-    def clut(self) -> list:
-        """Return the cLUT of the LUT16Type tag.
-
-        Returns:
-            list: The cLUT of the LUT16Type tag, a nested list structure
-                containing uInt16Number values.
-        """
-        if self._clut is not None:
-            return self._clut
-
-        # Calculate cLUT from tag data
-        i, o, g, n = self._i, self._o, self._g, self._n
-        tag_data = self._tagData
-        self._clut = [
-            [
-                [
-                    uInt16Number(
-                        tag_data[
-                            52 + n * i * 2 + o * 2 * (g * x + y) + z * 2 : 54
-                            + n * i * 2
-                            + o * 2 * (g * x + y)
-                            + z * 2
-                        ]
-                    )
-                    for z in range(o)
-                ]
-                for y in range(g)
-            ]
-            for x in range(int(g**i / g))
-        ]
-        return self._clut
-
-    @clut.setter
-    def clut(self, value: list) -> None:
-        """Set the cLUT of the LUT16Type tag.
-
-        Args:
-            value (list): The cLUT to set, a nested list structure containing
-                uInt16Number values.
-        """
-        self._clut = value
-
-    def clut_writepng(self, stream_or_filename: str | BinaryIO) -> None:
-        """Write the cLUT as a PNG image arranged in grid squares.
-
-        Args:
-            stream_or_filename (str | BinaryIO): The filename or
-                file-like object to write the PNG image to.
-
-        Raises:
-            NotImplementedError: If the output channels are not RGB
-                (3 channels).
-        """
-        if len(self.clut[0][0]) != 3:
-            raise NotImplementedError("clut_writepng: output channels != 3")
-        imfile.write(self.clut, stream_or_filename)
-
-    def clut_writecgats(self, stream_or_filename: str | BinaryIO) -> None:
-        """Write the cLUT as CGATS.
-
-        Args:
-            stream_or_filename (str | BinaryIO): The filename or
-                file-like object to write the CGATS data to.
-        """
-        # TODO:
-        # Need to take into account input/output curves
-        # Currently only supports RGB, A2B direction, and XYZ color space
-        if len(self.clut[0][0]) != 3:
-            raise NotImplementedError("clut_writecgats: output channels != 3")
-        if isinstance(stream_or_filename, str):
-            stream = open(stream_or_filename, "wb")  # noqa: SIM115
-        else:
-            stream = stream_or_filename
-        with stream:
-            stream.write(
-                b"""CTI3
-DEVICE_CLASS "DISPLAY"
-COLOR_REP "RGB_XYZ"
-BEGIN_DATA_FORMAT
-SAMPLE_ID RGB_R RGB_G RGB_B XYZ_X XYZ_Y XYZ_Z
-END_DATA_FORMAT
-BEGIN_DATA
-"""
-            )
-            clutres = len(self.clut[0])
-            block = 0
-            i = 1
-            if self.tagSignature and self.tagSignature.startswith("B2A"):
-                interp = [
-                    colormath.Interp(
-                        input_value, list(range(len(input_value))), use_numpy=True
-                    )
-                    for input_value in self.input
-                ]
-            for a in range(clutres):
-                for b in range(clutres):
-                    for c in range(clutres):
-                        R, G, B = [v / (clutres - 1.0) * 100 for v in (a, b, c)]  # noqa: N806
-                        if self.tagSignature and self.tagSignature.startswith("B2A"):
-                            linear_rgb = [
-                                interp[i](v)
-                                / (len(interp[i].xp) - 1.0)
-                                * (1 + (32767 / 32768.0))
-                                * 100
-                                for i, v in enumerate(self.clut[block][c])
-                            ]
-                            X, Y, Z = self.matrix.inverted() * linear_rgb  # noqa: N806
-                        else:
-                            X, Y, Z = [v / 32768.0 * 100 for v in self.clut[block][c]]  # noqa: N806
-                        stream.write(
-                            b"%i %7.3f %7.3f %7.3f %10.6f %10.6f %10.6f\n"
-                            % (i, R, G, B, X, Y, Z)
-                        )
-                        i += 1
-                    block += 1
-            stream.write(b"END_DATA\n")
-
-    @property
-    def clut_grid_steps(self) -> int:
-        """Return number of grid points per dimension.
-
-        Returns:
-            int: The number of grid points per dimension of the cLUT.
-        """
-        return self._g or len(self.clut[0])
-
-    @property
-    def input(self) -> list:
-        """Return the input table of the LUT16Type tag.
-
-        Returns:
-            list: The input table of the LUT16Type tag, a list of lists
-                containing uInt16Number values.
-        """
-        if self._input is None:
-            i, n = self._i, self._n
-            tag_data = self._tagData
-            self._input = [
-                [
-                    uInt16Number(
-                        tag_data[52 + n * 2 * z + y * 2 : 54 + n * 2 * z + y * 2]
-                    )
-                    for y in range(n)
-                ]
-                for z in range(i)
-            ]
-        return self._input
-
-    @input.setter
-    def input(self, value: list) -> None:
-        """Set the input table of the LUT16Type tag.
-
-        Args:
-            value (list): The input table to set, a list of lists containing
-                uInt16Number values.
-        """
-        self._input = value
-
-    @property
-    def input_channels_count(self) -> int:
-        """Return number of input channels.
-
-        Returns:
-            int: The number of input channels in the LUT16Type tag.
-        """
-        return self._i or len(self.input)
-
-    @property
-    def input_entries_count(self) -> int:
-        """Return number of entries per input channel.
-
-        Returns:
-            int: The number of entries per input channel in the LUT16Type tag.
-        """
-        return self._n or len(self.input[0])
-
-    def invert(self) -> None:
-        """Invert input and output tables."""
-        # Invert input/output 1d LUTs
-        for channel in (self.input, self.output):
-            for e, entries in enumerate(channel):
-                lut = {}
-                maxv = len(entries) - 1.0
-                for i, entry in enumerate(entries):
-                    lut[entry / 65535.0 * maxv] = i / maxv * 65535
-                xp = list(lut.keys())
-                fp = list(lut.values())
-                for i in range(len(entries)):
-                    if i not in lut:
-                        lut[i] = colormath.interp(i, xp, fp)
-                lut = dict_sort(lut)
-                channel[e] = list(lut.values())
-
-    def clut_row_apply_per_channel(
-        self,
-        indexes: list,
-        fn: Callable,
-        fnargs: None | tuple = None,
-        fnkwargs: None | dict = None,
-        pcs: None | str = None,
-        protect_gray_axis: bool = True,
-        protect_dark: bool = False,
-        protect_black: bool = True,
-        exclude: None | set = None,
-    ) -> None:
-        """Apply function to channel values of each cLUT row.
-
-        Args:
-            indexes (list): List of channel indexes to apply the function to.
-            fn (callable): Function to apply to the channel values.
-            fnargs (None | tuple, optional): Additional positional arguments to
-                pass to the function. Defaults to an empty tuple.
-            fnkwargs (None | dict, optional): Additional keyword arguments to
-                pass to the function. Defaults to an empty dictionary.
-            pcs (None | str, optional): The PCS (Profile Connection Space) to
-                use. Defaults to None, which means the PCS will be determined
-                from the profile.
-            protect_gray_axis (bool, optional): Whether to protect the gray
-                axis (diagonal) from modification. Defaults to True.
-            protect_dark (bool, optional): Whether to protect dark values from
-                modification. Defaults to False.
-            protect_black (bool, optional): Whether to protect black values
-                from modification. Defaults to True.
-            exclude (set, optional): Set of (row_index, column_index) tuples to
-                exclude from modification. Defaults to None, which means no
-                exclusions will be applied.
-        """
-        if fnargs is None:
-            fnargs = ()
-
-        if fnkwargs is None:
-            fnkwargs = {}
-
-        clutres = len(self.clut[0])
-        block = -1
-        for i, row in enumerate(self.clut):
-            channels = {}
-            for k in indexes:
-                channels[k] = []
-            if protect_gray_axis or protect_dark or protect_black or exclude:
-                if i % clutres == 0:
-                    block += 1
-                    gray_col_i = block if pcs == "XYZ" else clutres // 2  # L*a*b*
-                    gray_row_i = i + gray_col_i
-                fnkwargs["protect"] = []
-            for j, column in enumerate(row):
-                is_exclude = exclude and (i, j) in exclude
-                if is_exclude or (
-                    protect_gray_axis and (i == gray_row_i and j == gray_col_i)
-                ):
-                    if DEBUG:
-                        print(
-                            "protect", "exclude" if is_exclude else "gray", i, j, column
-                        )
-                    fnkwargs["protect"].append(j)
-                elif (protect_dark and sum(column) < 65535 * 0.03125 * 3) or (
-                    protect_black and min(column) == max(column) == 0
-                ):
-                    if DEBUG:
-                        print("protect dark", i, j, column)
-                    fnkwargs["protect"].append(j)
-                for k in indexes:
-                    channels[k].append(column[k])
-            for k in channels:
-                values = channels[k]
-                channels[k] = fn(values, *fnargs, **fnkwargs)
-            for j, column in enumerate(row):
-                for k in indexes:
-                    column[k] = channels[k][j]
-
-    def clut_shift_columns(
-        self,
-        order: tuple[int, int, int] = (1, 2, 0),
-    ) -> None:
-        """Shift cLUT columns, altering slowest to fastest changing column.
-
-        Args:
-            order (tuple[int, int, int]): The order of the channels to shift.
-                Default is (1, 2, 0), which means the first channel will be
-                the slowest changing, the second channel will be the middle
-                changing, and the third channel will be the fastest changing.
-
-        Raises:
-            NotImplementedError: If the number of input channels is not 3.
-        """
-        if len(self.input) != 3:
-            raise NotImplementedError("input channels != 3")
-        steps = len(self.clut[0])
-        clut = []
-        coord = [0, 0, 0]
-        for a in range(steps):
-            coord[order[0]] = a
-            for b in range(steps):
-                coord[order[1]] = b
-                clut.append([])
-                for c in range(steps):
-                    coord[order[2]] = c
-                    z, y, x = coord
-                    clut[-1].append(self.clut[z * steps + y][x])
-        self.clut = clut
-
-    @property
-    def matrix(self) -> colormath.Matrix3x3:
-        """Return the matrix of the LUT16Type tag.
-
-        Returns:
-            colormath.Matrix3x3: The matrix of the LUT16Type tag.
-        """
-        if self._matrix is None:
-            tag_data = self._tagData
-            return colormath.Matrix3x3(
-                [
-                    (
-                        s15Fixed16Number(tag_data[12:16]),
-                        s15Fixed16Number(tag_data[16:20]),
-                        s15Fixed16Number(tag_data[20:24]),
-                    ),
-                    (
-                        s15Fixed16Number(tag_data[24:28]),
-                        s15Fixed16Number(tag_data[28:32]),
-                        s15Fixed16Number(tag_data[32:36]),
-                    ),
-                    (
-                        s15Fixed16Number(tag_data[36:40]),
-                        s15Fixed16Number(tag_data[40:44]),
-                        s15Fixed16Number(tag_data[44:48]),
-                    ),
-                ]
-            )
-        return self._matrix
-
-    @matrix.setter
-    def matrix(self, value: colormath.Matrix3x3) -> None:
-        """Set the matrix of the LUT16Type tag.
-
-        Args:
-            value (colormath.Matrix3x3): The matrix to set.
-        """
-        self._matrix = value
-
-    @property
-    def output(self) -> list:
-        """Return the output table of the LUT16Type tag.
-
-        Returns:
-            list: The output table of the LUT16Type tag.
-        """
-        if self._output is None:
-            i, o, g, n, m = self._i, self._o, self._g, self._n, self._m
-            tag_data = self._tagData
-            self._output = [
-                [
-                    uInt16Number(
-                        tag_data[
-                            52 + n * i * 2 + m * 2 * z + y * 2 + g**i * o * 2 : 54
-                            + n * i * 2
-                            + m * 2 * z
-                            + y * 2
-                            + g**i * o * 2
-                        ]
-                    )
-                    for y in range(m)
-                ]
-                for z in range(o)
-            ]
-        return self._output
-
-    @output.setter
-    def output(self, value: list) -> None:
-        """Set the output table of the LUT16Type tag.
-
-        Args:
-            value (list): The output table to set, which should be a list of
-                lists containing uInt16Number values.
-        """
-        self._output = value
-
-    @property
-    def output_channels_count(self) -> int:
-        """Return number of output channels.
-
-        Returns:
-            int: The number of output channels of the cLUT.
-        """
-        return self._o or len(self.output)
-
-    @property
-    def output_entries_count(self) -> int:
-        """Return number of entries per output channel.
-
-        Returns:
-            int: The number of entries per output channel of the cLUT.
-        """
-        return self._m or len(self.output[0])
-
-    def smooth(
-        self,
-        diagpng: int = 2,
-        pcs: None | str = None,
-        filename: None | str = None,
-        logfile: None | TextIO = None,
-        debug_: int = 0,
-    ) -> None:
-        """Apply extra smoothing to the cLUT.
-
-        Args:
-            diagpng (int, optional): If 2, generate a diagnostic PNG image of
-                the cLUT. If 1, generate a diagnostic PNG image only if the
-                cLUT is modified. If 0, do not generate a diagnostic image.
-                Defatult is 2.
-            pcs (None | str, optional): The profile connection space, either
-                "XYZ" or "Lab". Default is None, which uses the profile's
-                connectionColorSpace if available.
-            filename (None | str, optional): The filename to save the
-                diagnostic image to. Default is None, which uses the
-                profile's filename if available.
-            logfile (None | TextIO, optional): A file-like object to write log
-                messages to. Default is None, which means no logging.
-            debug_ (int, optional): Debug level, where 0 means no debug, 1
-                means debug with some output, and 2 means full debug output.
-                Default is 0.
-
-        Raises:
-            TypeError: If PCS is not specified and no profile is available.
-        """
-        if not pcs:
-            if self.profile:
-                pcs = self.profile.connectionColorSpace
-            else:
-                raise TypeError("PCS not specified")
-
-        if not filename and self.profile:
-            filename = self.profile.filename
-
-        clutres = len(self.clut[0])
-
-        sig = self.tagSignature or id(self)
-
-        if diagpng and filename and len(self.output) == 3:
-            # Generate diagnostic images
-            fname, _ = os.path.splitext(filename)
-            diag_fname = f"{fname}.{sig}.post.CLUT.png"
-            if diagpng == 2 and not os.path.isfile(diag_fname):
-                self.clut_writepng(diag_fname)
-        else:
-            diagpng = 0
-
-        if logfile:
-            logfile.write(f"Smoothing {sig}...\n")
-        # Create a list of <clutres> number of 2D grids, each one with a
-        # size of (width x height) <clutres> x <clutres>
-        grids = []
-        for i, block in enumerate(self.clut):
-            if i % clutres == 0:
-                grids.append([])
-            grids[-1].append([])
-            for RGB in block:  # noqa: N806
-                grids[-1][-1].append(RGB)
-        for i, grid in enumerate(grids):
-            for y in range(clutres):
-                for x in range(clutres):
-                    is_dark = sum(grid[y][x]) < 65535 * 0.03125 * 3
-                    if pcs == "XYZ":
-                        is_gray = x == y == i
-                    elif clutres // 2 != clutres / 2.0:
-                        # For CIELab cLUT, gray will only
-                        # fall on a cLUT point if uneven cLUT res
-                        is_gray = x == y == clutres // 2
-                    else:
-                        is_gray = False
-                    # print(
-                    #     i, y, x,
-                    #     "{:d} {:d} {:d}".format(*(int(v / 655.35 * 2.55)
-                    #     for v in grid[y][x])),
-                    #     is_dark,
-                    #     raw_input(is_gray) if is_gray else "",
-                    # )
-                    if is_dark or is_gray:
-                        # Don't smooth dark colors and gray axis
-                        continue
-                    RGB = [[v] for v in grid[y][x]]  # noqa: N806
-                    # Use either "plus"-shaped or box filter depending if one
-                    # channel is fully saturated
-                    if clutres - 1 in (y, x) or 0 in (x, y):
-                        # Filter with a "plus" (+) shape
-                        # Smoothing factor for L*a*b* -> RGB cLUT above 50%
-                        smooth = 0.25 if pcs == "Lab" and i > clutres / 2.0 else 0.5
-                        for j, c in enumerate((x, y)):
-                            # Omit corners and perpendicular axis
-                            if 0 < c < clutres - 1:
-                                for n in (-1, 1):
-                                    yi, xi = (y, y + n)[j], (x + n, x)[j]
-                                    if -1 < xi < clutres and -1 < yi < clutres:
-                                        RGBn = grid[yi][xi]  # noqa: N806
-                                        if debug_ == 2:
-                                            if i < clutres - 1 or grid[y][x] != [
-                                                16384,
-                                                16384,
-                                                16384,
-                                            ]:
-                                                grid[y][x] = [32768, 32768, 32768]
-                                            if x == y == clutres - 2:
-                                                RGBn[:] = [16384, 16384, 16384]
-                                        for k in range(3):
-                                            RGB[k].append(
-                                                RGBn[k] * smooth
-                                                + RGB[k][0] * (1 - smooth)
-                                            )
-                    else:
-                        # Box filter, 3x3
-                        # Center pixel weight = 1.0, surround = 2/3, corners = 1/3
-                        if debug_ == 1:
-                            grid[y][x] = [32768, 32768, 32768]
-                        for j in (0, 1):
-                            for n in (-1, 1):
-                                for yi, xi in [
-                                    ((y, y + n)[j], (x + n, x)[j]),
-                                    (y - n, (x + n, x - n)[j]),
-                                ]:
-                                    if -1 < xi < clutres and -1 < yi < clutres:
-                                        RGBn = grid[yi][xi]  # noqa: N806
-                                        if yi != y and xi != x:
-                                            smooth = 1 / 3.0
-                                        else:
-                                            smooth = 2 / 3.0
-                                        if debug_ == 1 and x == y == clutres - 2:
-                                            RGBn[:] = (v * (1 - smooth) for v in RGBn)
-                                        for k in range(3):
-                                            RGB[k].append(
-                                                RGBn[k] * smooth
-                                                + RGB[k][0] * (1 - smooth)
-                                            )
-                    if not debug_:
-                        grid[y][x] = [sum(v) / float(len(v)) for v in RGB]
-            for j, row in enumerate(grid):
-                self.clut[i * clutres + j] = [
-                    [min(v, 65535) for v in RGB] for RGB in row
-                ]
-
-        if diagpng and filename:
-            self.clut_writepng(f"{fname}.{sig}.post.CLUT.smooth.png")
-
-    def smooth2(
-        self,
-        diagpng: int = 2,
-        pcs: None | str = None,
-        filename: None | str = None,
-        logfile: None | TextIO = None,
-        window: tuple[float, float, float] = (1 / 16.0, 1, 1 / 16.0),
-    ) -> None:
-        """Apply extra smoothing to the cLUT.
-
-        Args:
-            diagpng (int, optional): Diagnostic PNG generation level (0, 1, 2,
-                or 3). If `diagpng` is 0, no diagnostic images will be
-                generated. If `diagpng` is 1, only the final smoothed cLUT will
-                be saved. If `diagpng` is 2, the original and final smoothed
-                cLUT will be saved. If `diagpng` is 3, intermediate steps will
-                also be saved. If `diagpng` is 2 or 3, the filename will be
-                used to generate diagnostic PNG files with the signature of the
-                tag appended to the filename, e.g.,
-                `filename.<signature>.post.CLUT.png`.
-            pcs (None | str): The profile connection space (PCS) to use, e.g.,
-                "Lab" or "XYZ". Default is None, which will use the PCS from
-                the profile if available, or raise a TypeError if no PCS is
-                specified and no profile is available.
-            filename (None | str): The filename to save diagnostic images to.
-                Default is None, which will use the profile filename if
-                available.
-            logfile (None | TextIO): A file-like object to write log messages
-                to. Default is None, which means no logging will be done.
-            window (tuple[float, float, float]): The smoothing window as a
-                tuple of three floats, representing the weights for the R, G,
-                and B channels. Default value is (1/16.0, 1, 1/16.0).
-
-        Raises:
-            TypeError: If PCS is not specified and no profile is available.
-        """
-        if not pcs:
-            if self.profile:
-                pcs = self.profile.connectionColorSpace
-            else:
-                raise TypeError("PCS not specified")
-
-        if not filename and self.profile:
-            filename = self.profile.filename
-
-        clutres = len(self.clut[0])
-
-        sig = self.tagSignature or id(self)
-
-        if diagpng and filename and len(self.output) == 3:
-            # Generate diagnostic images
-            fname, ext = os.path.splitext(filename)
-            diag_fname = f"{fname}.{sig}.post.CLUT.png"
-            if diagpng == 2 and not os.path.isfile(diag_fname):
-                self.clut_writepng(diag_fname)
-        else:
-            diagpng = 0
-
-        if logfile:
-            logfile.write(f"Smoothing {sig}...\n")
-
-        for i in range(3):
-            state = ("original", "pass", "final")[i]
-            if diagpng != 3 and i != 1:
-                continue
-            for j, (order, channels) in enumerate(
-                [
-                    (None, "BGR"),
-                    ((1, 2, 0), "RBG"),
-                    ((0, 2, 1), "BRG"),
-                    ((2, 1, 0), "GRB"),
-                    ((0, 2, 1), "RGB"),
-                    ((2, 0, 1), "GBR"),
-                    ((0, 2, 1), "BGR"),
-                ]
-            ):
-                if order:
-                    if DEBUG:
-                        print("Shifting order to", channels)
-                    self.clut_shift_columns(order)
-                if i == 1 and j != 6:
-                    if DEBUG:
-                        print("Smoothing")
-                    exclude = None
-                    protect_gray_axis = True
-                    if pcs == "Lab":
-                        if clutres // 2 != clutres / 2.0:
-                            # For CIELab cLUT, gray will only
-                            # fall on a cLUT point if uneven cLUT res
-                            if channels in ("RBG", "RGB"):
-                                exclude = [
-                                    ((clutres // 2 + 1) * (clutres - 1), col)
-                                    for col in range(clutres)
-                                ]
-                                protect_gray_axis = False
-                            elif channels in ("BRG", "GRB"):
-                                exclude = [
-                                    ((clutres // 2) * clutres + y, clutres // 2)
-                                    for y in range(clutres)
-                                ]
-                                protect_gray_axis = False
-                        else:
-                            protect_gray_axis = False
-                    self.clut_row_apply_per_channel(
-                        (0, 1, 2),
-                        colormath.smooth_avg,
-                        (),
-                        {"window": window},
-                        pcs,
-                        protect_gray_axis,
-                        exclude=exclude,
-                    )
-                if diagpng == 3 and filename and j != 6:
-                    if DEBUG:
-                        print("Writing diagnostic PNG for", state, channels)
-                    self.clut_writepng(
-                        f"{fname}.{sig}.post.CLUT.{channels}.{state}.png"
-                    )
-
-        if diagpng and filename:
-            self.clut_writepng(f"{fname}.{sig}.post.CLUT.smooth.png")
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: Raw tag data for the LUT16Type tag.
-        """
-        if (self._matrix, self._input, self._clut, self._output) == (None,) * 4:
-            return self._tagData
-        tag_data = [
-            b"mft2",
-            b"\0" * 4,
-            uInt8Number_tohex(len(self.input)),
-            uInt8Number_tohex(len(self.output)),
-            uInt8Number_tohex(len(self.clut and self.clut[0])),
-            b"\0",
-            s15Fixed16Number_tohex(self.matrix[0][0]),
-            s15Fixed16Number_tohex(self.matrix[0][1]),
-            s15Fixed16Number_tohex(self.matrix[0][2]),
-            s15Fixed16Number_tohex(self.matrix[1][0]),
-            s15Fixed16Number_tohex(self.matrix[1][1]),
-            s15Fixed16Number_tohex(self.matrix[1][2]),
-            s15Fixed16Number_tohex(self.matrix[2][0]),
-            s15Fixed16Number_tohex(self.matrix[2][1]),
-            s15Fixed16Number_tohex(self.matrix[2][2]),
-            uInt16Number_tohex(len(self.input and self.input[0])),
-            uInt16Number_tohex(len(self.output and self.output[0])),
-        ]
-        for entries in self.input:
-            tag_data.extend(uInt16Number_tohex(v) for v in entries)
-        for block in self.clut:
-            for entries in block:
-                tag_data.extend(uInt16Number_tohex(v) for v in entries)
-        for entries in self.output:
-            tag_data.extend(uInt16Number_tohex(v) for v in entries)
-        return b"".join(tag_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set raw tag data.
-
-        Args:
-            tagData (bytes): Raw tag data.
-        """
-        self._tagData = tagData
-
-
 class Observer(ADict):
     """ICC Observer tag.
 
@@ -4210,833 +3305,6 @@ class ColorantTableType(ICCProfileTag, AODict):
             name = data[:end]
             self[name] = AODict(list(zip(keys, pcsvalues)))
             data = data[38:]
-
-
-class CurveType(ICCProfileTag, list):
-    """ICC CurveType tag.
-
-    Args:
-        tagData (bytes, optional): Raw tag data. Defaults to None.
-        tagSignature (str, optional): Tag signature. Defaults to None.
-        profile (ICCProfile, optional): ICC profile instance. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        tagData: None | bytes = None,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-        profile: None | ICCProfile = None,
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        self.profile = profile
-        self._reset()
-        if not tagData:
-            return
-        curve_entries_count = uInt32Number(tagData[8:12])
-        curve_entries = tagData[12:]
-        if curve_entries_count == 1:
-            # Gamma
-            self.append(u8Fixed8Number(curve_entries[:2]))
-        elif curve_entries_count:
-            # Curve
-            for _count in range(curve_entries_count):
-                self.append(uInt16Number(curve_entries[:2]))
-                curve_entries = curve_entries[2:]
-        else:
-            # Identity
-            self.append(1.0)
-
-    def __delitem__(self, value: int) -> None:
-        """Delete an item from the list.
-
-        Args:
-            value (int): Index of the item to delete.
-        """
-        list.__delitem__(self, value)
-        self._reset()
-
-    def __iadd__(self, value: Any) -> Self:  # noqa: ANN401
-        """Add a value to the list.
-
-        Args:
-            value (Any): The value to add to the list.
-
-        Returns:
-            Self: The updated list.
-        """
-        list.__iadd__(self, value)
-        self._reset()
-        return self
-
-    def __imul__(self, value: int) -> Self:
-        """Multiply the list by a scalar.
-
-        Args:
-            value (int): The scalar to multiply the list by.
-
-        Returns:
-            Self: The updated list.
-        """
-        list.__imul__(self, value)
-        self._reset()
-        return self
-
-    def __setitem__(self, key: int, value: Any) -> None:  # noqa: ANN401
-        """Set an item in the list.
-
-        Args:
-            key (int): Index of the item to set.
-            value (Any): The new value to set at the index.
-        """
-        list.__setitem__(self, key, value)
-        self._reset()
-
-    def _reset(self) -> None:
-        """Reset internal state."""
-        self._transfer_function = {}
-        self._bt1886 = {}
-
-    def append(self, object_: Any) -> None:  # noqa: ANN401
-        """Append an object to the list.
-
-        Args:
-            object_ (Any): The object to append to the list.
-        """
-        list.append(self, object_)
-        self._reset()
-
-    def apply_bpc(self, black_Y_out: float = 0, weight: bool = False) -> None:  # noqa: N803
-        """Apply black point compensation to the curve.
-
-        Args:
-            black_Y_out (float, optional): The output black point Y value to
-                apply. Defaults to 0.
-            weight (bool, optional): If True, apply weighted black point
-                compensation. Defaults to False.
-        """
-        if len(self) < 2:
-            return
-        D50_xyY = colormath.XYZ2xyY(*colormath.get_whitepoint("D50"))  # noqa: N806
-        bp_in = colormath.xyY2XYZ(D50_xyY[0], D50_xyY[1], self[0] / 65535.0)
-        bp_out = colormath.xyY2XYZ(D50_xyY[0], D50_xyY[1], black_Y_out)
-        wp_out = colormath.xyY2XYZ(D50_xyY[0], D50_xyY[1], self[-1] / 65535.0)
-        for i, v in enumerate(self):
-            X, Y, Z = colormath.xyY2XYZ(D50_xyY[0], D50_xyY[1], v / 65535.0)  # noqa: N806
-            self[i] = (
-                colormath.apply_bpc(X, Y, Z, bp_in, bp_out, wp_out, weight)[1] * 65535.0
-            )
-
-    def extend(self, iterable: Iterable) -> None:
-        """Extend the list with elements from an iterable.
-
-        Args:
-            iterable (iterable): An iterable whose elements will be added to
-                the list.
-        """
-        list.extend(self, iterable)
-        self._reset()
-
-    def get_gamma(
-        self,
-        use_vmin_vmax: bool = False,
-        average: bool = True,
-        least_squares: bool = False,
-        slice_: tuple[float, float] = (0.01, 0.99),
-        lstar_slice: bool = True,
-    ) -> float | list:
-        """Return average or least squares gamma or a list of gamma values.
-
-        Args:
-            use_vmin_vmax (bool, optional): If True, use the first and last
-                values as vmin and vmax for gamma calculation. Default is
-                False.
-            average (bool, optional): If True, return the average gamma value.
-                Default is True.
-            least_squares (bool, optional): If True, return the least squares
-                gamma value. Default is False.
-            slice_ (tuple[float, float], optional): The range of the curve to
-                consider for the gamma calculation. Defaults to (0.01, 0.99).
-            lstar_slice (bool, optional): If True, use L* values for the slice.
-                Defaults to True.
-
-        Returns:
-            float | list: If average or least_squares is True, return a single
-                gamma value. Otherwise, return a list of gamma values for the
-                specified slice.
-        """
-        if len(self) <= 1:
-            values = self if len(self) else [1.0]  # Identity
-            if average or least_squares:
-                return values[0]
-            return [values[0]]
-        if lstar_slice:
-            start = slice_[0] * 100
-            end = slice_[1] * 100
-            values = []
-            for i, y in enumerate(self):
-                n = colormath.XYZ2Lab(0, y / 65535.0 * 100, 0)[0]
-                if start <= n <= end:
-                    values.append((i / (len(self) - 1.0) * 65535.0, y))
-        else:
-            maxv = len(self) - 1.0
-            maxi = int(maxv)
-            starti = round(slice_[0] * maxi)
-            endi = round(slice_[1] * maxi) + 1
-            values = list(
-                zip(
-                    [(v / maxv) * 65535 for v in range(starti, endi)], self[starti:endi]
-                )
-            )
-        vmin = 0
-        vmax = 65535.0
-        if use_vmin_vmax and len(self) > 2:
-            vmin = self[0]
-            vmax = self[-1]
-        return colormath.get_gamma(values, 65535.0, vmin, vmax, average, least_squares)
-
-    def get_transfer_function(
-        self,
-        best: bool = True,
-        slice_: tuple[float, float] = (0.05, 0.95),
-        black_Y: None | float = None,  # noqa: N803
-        outoffset: None | float = None,
-    ) -> tuple[tuple[str, float, float], float] | float:
-        """Return transfer function name, exponent and match percentage.
-
-        Args:
-            best (bool): If True, return the best matching transfer function.
-            slice_ (tuple[float, float]): The range of the curve to consider for
-                the transfer function.
-            black_Y (None | float): The black Y value to use for the transfer
-                function calculation. If None, it will be calculated from the
-                curve.
-            outoffset (None | float): The output offset to apply. If None, it
-                will be set to 1.0.
-
-        Returns:
-            tuple: A tuple containing the transfer function name, exponent,
-                output offset, and the match percentage.
-            float: The match percentage of the transfer function.
-        """
-        if len(self) == 1:
-            # Gamma
-            return (f"Gamma {round(self[0], 2):.2f}", self[0], 1.0), 1.0
-        if not len(self):
-            # Identity
-            return ("Gamma 1.0", 1.0, 1.0), 1.0
-        transfer_function = self._transfer_function.get((best, slice_))
-        if transfer_function:
-            return transfer_function
-        trc = CurveType()
-        match = {}
-        otrc = CurveType()
-        otrc[:] = self
-        if otrc[0]:
-            otrc.apply_bpc()
-        vmin = otrc[0]
-        vmax = otrc[-1]
-        if self.profile and isinstance(self.profile.tags.get("lumi"), XYZType):
-            white_cdm2 = self.profile.tags.lumi.Y
-        else:
-            white_cdm2 = 100.0
-        if black_Y is None:
-            black_Y = self[0] / 65535.0  # noqa: N806
-        black_cdm2 = black_Y * white_cdm2
-        maxv = len(otrc) - 1.0
-        maxi = int(maxv)
-        _starti = round(0.4 * maxi)
-        _endi = round(0.6 * maxi)
-        gamma = otrc.get_gamma(True, slice_=(0.4, 0.6), lstar_slice=False)
-        egamma = colormath.get_gamma([(0.5, 0.5**gamma)], vmin=-black_Y)
-        outoffset_unspecified = outoffset is None
-        if outoffset_unspecified:
-            outoffset = 1.0
-        tfs = [
-            ("Rec. 709", -709, outoffset),
-            ("Rec. 1886", -1886, 0),
-            ("SMPTE 240M", -240, outoffset),
-            ("SMPTE 2084", -2084, outoffset),
-            ("DICOM", -1023, outoffset),
-            ("HLG", -2, outoffset),
-            ("L*", -3.0, outoffset),
-            ("sRGB", -2.4, outoffset),
-            (
-                f"Gamma {gamma:.2f} {outoffset:.0%}",
-                gamma,
-                outoffset,
-            ),
-        ]
-        if outoffset_unspecified and black_Y:
-            tfs.extend(
-                (
-                    f"Gamma {gamma:.2f} {i:d}%",
-                    gamma,
-                    i / 100.0,
-                )
-                for i in range(100)
-            )
-        for name, exp, outoffset in tfs:
-            if name in ("DICOM", "Rec. 1886", "SMPTE 2084", "HLG"):
-                try:
-                    if name == "DICOM":
-                        trc.set_dicom_trc(black_cdm2, white_cdm2, size=len(self))
-                    elif name == "Rec. 1886":
-                        trc.set_bt1886_trc(black_Y, size=len(self))
-                    elif name == "SMPTE 2084":
-                        trc.set_smpte2084_trc(black_cdm2, white_cdm2, size=len(self))
-                    elif name == "HLG":
-                        trc.set_hlg_trc(black_cdm2, white_cdm2, size=len(self))
-                except ValueError:
-                    continue
-            elif exp > 0 and black_Y:
-                trc.set_bt1886_trc(black_Y, outoffset, egamma, "b")
-            else:
-                trc.set_trc(exp, len(self), vmin, vmax)
-            if trc[0] and trc[-1] - trc[0]:
-                trc.apply_bpc()
-            if otrc == trc:
-                match[(name, exp, outoffset)] = 1.0
-            else:
-                match[(name, exp, outoffset)] = 0.0
-                count = 0
-                start = slice_[0] * len(self)
-                end = slice_[1] * len(self)
-                for i, n in enumerate(otrc):
-                    # n = colormath.XYZ2Lab(0, n / 65535.0 * 100, 0)[0]
-                    if start > i or i > end:
-                        continue
-                    n = colormath.get_gamma(
-                        [(i / (len(self) - 1.0) * 65535.0, n)],
-                        65535.0,
-                        vmin,
-                        vmax,
-                        False,
-                    )
-                    if n:
-                        n = n[0]
-                        # n2 = colormath.XYZ2Lab(0, trc[i] / 65535.0 * 100, 0)[0]
-                        n2 = colormath.get_gamma(
-                            [(i / (len(self) - 1.0) * 65535.0, trc[i])],
-                            65535.0,
-                            vmin,
-                            vmax,
-                            False,
-                        )
-                        if n2 and n2[0]:
-                            n2 = n2[0]
-                            match[(name, exp, outoffset)] += 1 - (
-                                max(n, n2) - min(n, n2)
-                            ) / ((n + n2) / 2.0)
-                            count += 1
-                if count:
-                    match[(name, exp, outoffset)] /= count
-        if not best:
-            self._transfer_function[(best, slice_)] = match
-            return match
-        match, (name, exp, outoffset) = sorted(
-            zip(list(match.values()), list(match.keys()))
-        )[-1]
-        self._transfer_function[(best, slice_)] = (name, exp, outoffset), match
-        return (name, exp, outoffset), match
-
-    def insert(self, object_: Any) -> None:  # noqa: ANN401
-        """Insert an item at a given position in the list.
-
-        Args:
-            object_ (Any): The item to insert into the list.
-        """
-        list.insert(self, object_)
-        self._reset()
-
-    def pop(self, index: int) -> None:
-        """Remove and return an item at the given index.
-
-        Args:
-            index (int): The index of the item to remove and return.
-        """
-        list.pop(self, index)
-        self._reset()
-
-    def remove(self, value: Any) -> None:  # noqa: ANN401
-        """Remove the first occurrence of a value from the list.
-
-        Args:
-            value (Any): The value to remove from the list.
-        """
-        list.remove(self, value)
-        self._reset()
-
-    def reverse(self) -> None:
-        """Reverse the order of the list."""
-        list.reverse(self)
-        self._reset()
-
-    def set_bt1886_trc(
-        self,
-        black_Y: float = 0,  # noqa: N803
-        outoffset: float = 0.0,
-        gamma: float = 2.4,
-        gamma_type: str = "B",
-        size: None | int = None,
-    ) -> None | colormath.BT1886:
-        """Set the response to the BT. 1886 curve.
-
-        This response is special in that it depends on the actual black
-        level of the display.
-
-        Args:
-            black_Y (float, optional): Black point in absolute Y, range 0..100.
-                Defaults to 0.
-            outoffset (float, optional): Output offset, range 0.0..1. Defaults to 0.0.
-            gamma (float, optional): Gamma value, range 1.0..3.0.
-                Defaults to 2.4.
-            gamma_type (str, optional): Type of gamma to use, either "b" for
-                BT.1886 or "g" for technical gamma. Defaults to "B".
-            size (None | int, optional): Number of steps. Recommended >= 1024.
-
-        Returns:
-            None | BT1886: None if the response was set successfully, or BT1886
-                instance if it was already set for the given parameters.
-        """
-        bt1886 = self._bt1886.get((gamma, black_Y, outoffset))
-        if bt1886:
-            return bt1886
-        if gamma_type in ("b", "g"):
-            # Get technical gamma needed to achieve effective gamma
-            gamma = colormath.xicc_tech_gamma(gamma, black_Y, outoffset)
-        rXYZ = colormath.RGB2XYZ(1.0, 0, 0)  # noqa: N806
-        gXYZ = colormath.RGB2XYZ(0, 1.0, 0)  # noqa: N806
-        bXYZ = colormath.RGB2XYZ(0, 0, 1.0)  # noqa: N806
-        mtx = colormath.Matrix3x3(
-            [
-                [rXYZ[0], gXYZ[0], bXYZ[0]],
-                [rXYZ[1], gXYZ[1], bXYZ[1]],
-                [rXYZ[2], gXYZ[2], bXYZ[2]],
-            ]
-        )
-        wXYZ = colormath.RGB2XYZ(1.0, 1.0, 1.0)  # noqa: N806
-        x, y = colormath.XYZ2xyY(*wXYZ)[:2]
-        XYZbp = colormath.xyY2XYZ(x, y, black_Y)  # noqa: N806
-        bt1886 = colormath.BT1886(mtx, XYZbp, outoffset, gamma)
-        self._bt1886[(gamma, black_Y, outoffset)] = bt1886
-        self.set_trc(-709, size)
-        for i, v in enumerate(self):
-            X, Y, Z = colormath.xyY2XYZ(x, y, v / 65535.0)  # noqa: N806
-            self[i] = bt1886.apply(X, Y, Z)[1] * 65535.0
-        return None
-
-    def set_dicom_trc(
-        self, black_cdm2: float = 0.05, white_cdm2: float = 100, size: None | int = None
-    ) -> None:
-        """Set the response to the DICOM Grayscale Standard Display Function.
-
-        This response is special in that it depends on the actual black
-        and white level of the display.
-
-        Args:
-            black_cdm2 (float, optional): Black point in absolute Y,
-                range 0.05..white_cdm2. Defaults to 0.05.
-            white_cdm2 (float, optional): White point in absolute Y,
-                range black_cdm2..4000. Defaults to 100.
-            size (None | int, optional): Number of steps. Recommended >= 1024.
-        """
-        # See http://medical.nema.org/Dicom/2011/11_14pu.pdf
-        # Luminance levels depend on the start level of 0.05 cd/m2
-        # and end level of 4000 cd/m2
-        black_cdm2 = round(black_cdm2, 6)
-        if black_cdm2 < 0.05 or black_cdm2 >= white_cdm2:
-            raise ValueError(
-                f"The black level of {black_cdm2} cd/m2 is out of range "
-                "for DICOM. Valid range begins at 0.05 cd/m2."
-            )
-        if white_cdm2 > 4000 or white_cdm2 <= black_cdm2:
-            raise ValueError(
-                f"The white level of {white_cdm2} cd/m2 is out of range "
-                "for DICOM. Valid range is up to 4000 cd/m2."
-            )
-        black_jndi = colormath.DICOM(black_cdm2, True)
-        white_jndi = colormath.DICOM(white_cdm2, True)
-        white_dicom_y = math.pow(10, colormath.DICOM(white_jndi))
-        if not size:
-            size = len(self)
-        if size < 2:
-            size = 1024
-        self[:] = []
-        for i in range(size):
-            v = (
-                math.pow(
-                    10,
-                    colormath.DICOM(
-                        black_jndi + (float(i) / (size - 1)) * (white_jndi - black_jndi)
-                    ),
-                )
-                / white_dicom_y
-            )
-            self.append(v * 65535)
-
-    def set_hlg_trc(
-        self,
-        black_cdm2: float = 0,
-        white_cdm2: float = 100,
-        system_gamma: float = 1.2,
-        ambient_cdm2: float = 5,
-        maxsignal: float = 1.0,
-        size: None | int = None,
-        logfile: None | TextIO = None,
-    ) -> None:
-        """Set the response to the Hybrid Log-Gamma (HLG) function.
-
-        This response is special in that it depends on the actual black
-        and white level of the display, system gamma and ambient.
-
-        Args:
-            black_cdm2 (float, optional): Black point in absolute XYZ, range
-                0..white_cdm2. Defaults to 0.
-            white_cdm2 (float, optional): White point in absolute Y,
-                range 0..10000. Defaults to 100.
-            system_gamma (float, optional): System gamma, typically 1.2.
-                Defaults to 1.2.
-            ambient_cdm2 (float, optional): Ambient light in cd/m2. Defaults to
-                5.
-            maxsignal (float, optional): Set clipping point. Defaults to 1.0.
-            size (None | int, optional): Number of steps. Recommended >= 1024.
-            logfile (None | TextIO, optional): Log file to write diagnostic
-                information to. Defaults to None.
-
-        Raises:
-            ValueError: If the black or white levels are out of range for
-                HLG.
-            ValueError: If the white level exceeds 10000 cd/m2.
-            ValueError: If the black level is negative or greater than or equal
-                to the white level.
-            ValueError: If the white level is less than or equal to the black
-                level.
-        """
-        if black_cdm2 < 0 or black_cdm2 >= white_cdm2:
-            raise ValueError(
-                f"The black level of {black_cdm2:f} cd/m2 is out of range "
-                "for HLG. Valid range begins at 0 cd/m2."
-            )
-        values = []
-
-        hlg = colormath.HLG(black_cdm2, white_cdm2, system_gamma, ambient_cdm2)
-
-        if maxsignal < 1:
-            # Adjust EOTF so that EOTF[maxsignal] gives (approx) white_cdm2
-            while hlg.eotf(maxsignal) * hlg.white_cdm2 < white_cdm2:
-                hlg.white_cdm2 += 1
-
-        lscale = 1.0 / hlg.oetf(1.0, True)
-        hlg.white_cdm2 *= lscale
-        if lscale < 1 and logfile:
-            logfile.write(
-                f"Nominal peak luminance after scaling = {hlg.white_cdm2:.2f}\n"
-            )
-
-        maxv = hlg.eotf(maxsignal)
-        if not size:
-            size = len(self)
-        if size < 2:
-            size = 1024
-        for i in range(size):
-            n = i / (size - 1.0)
-            v = hlg.eotf(min(n, maxsignal))
-            values.append(min(v / maxv, 1.0))
-        self[:] = [min(v * 65535, 65535) for v in values]
-
-    def set_smpte2084_trc(
-        self,
-        black_cdm2: float = 0,
-        white_cdm2: float = 100,
-        master_black_cdm2: float = 0,
-        master_white_cdm2: float = 0,
-        use_alternate_master_white_clip: bool = True,
-        rolloff: bool = False,
-        size: None | int = None,
-    ) -> None:
-        """Set the response to the SMPTE 2084 perceptual quantizer (PQ) function.
-
-        This response is special in that it depends on the actual black
-        and white level of the display.
-
-        Args:
-            black_cdm2 (float, optinoal): Black point in absolute Y, range
-                0..white_cdm2. Defaults to 0.
-            white_cdm2 (float, optional): White point in absolute Y, range
-                0..10000. Defaults to 100.
-            master_black_cdm2 (float, optional): Used to normalize PQ values.
-            master_white_cdm2 (float, optional): Used to normalize PQ values.
-            use_alternate_master_white_clip (bool, optional): If True,
-                use the alternate master white clip as defined in ITU-R
-                BT.2390. Defaults to True.
-            rolloff (bool, optional): BT.2390.
-            size (None | int, optional): Number of steps. Recommended >= 1024.
-
-        Raises:
-            ValueError: If the black or white levels are out of range for
-                SMPTE 2084.
-            ValueError: If the white level exceeds 10000 cd/m2.
-            ValueError: If the black level is negative or greater than or equal
-                to the white level.
-            ValueError: If the white level is less than or equal to the black
-                level.
-            ValueError: If the master white level exceeds 10000 cd/m2.
-        """
-        # See https://www.smpte.org/sites/default/files/2014-05-06-EOTF-Miller-1-2-handout.pdf
-        # Luminance levels depend on the end level of 10000 cd/m2
-        if black_cdm2 < 0 or black_cdm2 >= white_cdm2:
-            raise ValueError(
-                f"The black level of {black_cdm2:f} cd/m2 is out of range "
-                "for SMPTE 2084. Valid range begins at 0 cd/m2."
-            )
-        if max(white_cdm2, master_white_cdm2) > 10000:
-            raise ValueError(
-                f"The white level of {max(white_cdm2, master_white_cdm2):f} "
-                "cd/m2 is out of range for SMPTE 2084. "
-                "Valid range is up to 10000 cd/m2."
-            )
-        values = []
-        maxv = white_cdm2 / 10000.0
-        maxi = colormath.special_pow(maxv, 1.0 / -2084)
-        if rolloff:
-            # Rolloff as defined in ITU-R BT.2390
-            if not master_white_cdm2:
-                master_white_cdm2 = 10000
-            bt2390 = colormath.BT2390(
-                black_cdm2,
-                white_cdm2,
-                master_black_cdm2,
-                master_white_cdm2,
-                use_alternate_master_white_clip,
-            )
-            maxi_out = maxi
-        else:
-            if not master_white_cdm2:
-                master_white_cdm2 = white_cdm2
-            maxi_out = colormath.special_pow(master_white_cdm2 / 10000.0, 1.0 / -2084)
-        if not size:
-            size = len(self)
-        if size < 2:
-            size = 1024
-        for i in range(size):
-            n = i / (size - 1.0)
-            if rolloff:
-                n = bt2390.apply(n)
-            v = colormath.special_pow(n * (maxi / maxi_out), -2084)
-            values.append(min(v / maxv, 1.0))
-        self[:] = [min(v * 65535, 65535) for v in values]
-        if black_cdm2 and not rolloff:
-            self.apply_bpc(black_cdm2 / white_cdm2)
-
-    def set_trc(
-        self,
-        power: float | Callable = 2.2,
-        size: None | int = None,
-        vmin: float = 0,
-        vmax: float = 65535,
-    ) -> None:
-        """Set the response to a certain function.
-
-        Args:
-            power (float | Callable, optional): The power to raise the input
-                value to. If a callable, it should take a single float argument
-                and return a float. Defaults to 2.2. Positive power, or
-                -2.4 = sRGB, -3.0 = L*, -240 = SMPTE 240M, -601 = Rec. 601,
-                -709 = Rec. 709 (Rec. 601 and 709 transfer functions are
-                identical).
-            size (int, optional): The number of entries in the curve. Defaults
-                to None.
-            vmin (float, optional): The minimum value of the curve. Defaults to
-                0.
-            vmax (float, optional): The maximum value of the curve. Defaults to
-                65535.
-        """
-        if not size:
-            size = len(self) or 1024
-        if size == 1:
-            if callable(power):
-                power = colormath.get_gamma([(0.5, power(0.5))])
-            if power >= 0.0 and not vmin:
-                self[:] = [power]
-                return
-            size = 1024
-        self[:] = []
-        if not callable(power):
-            exp = power
-
-            def power(a: float) -> float:
-                """Power function for non-callable power.
-
-                Args:
-                    a (float): The input value to raise to the power.
-
-                Returns:
-                    float: The result of raising the input value to the power.
-                """
-                return colormath.special_pow(a, exp)
-
-        for i in range(size):
-            self.append(vmin + power(float(i) / (size - 1)) * (vmax - vmin))
-
-    def smooth_cr(self, length: int = 64) -> None:
-        """Smooth curves (Catmull-Rom).
-
-        Args:
-            length (int, optional): Number of points to use for smoothing.
-                Defaults to 64.
-        """
-        raise NotImplementedError
-
-    def smooth_avg(
-        self,
-        passes: int = 1,
-        window: None | tuple[float, float, float] = None,
-    ) -> None:
-        """Smooth curves (moving average).
-
-        Args:
-            passes (int, optional): Number of passes. Defaults to 1.
-            window (None | tuple[float, float, float], optional): Tuple or list
-                containing weighting factors. Its length determines the size of
-                the window to use. Defaults to (1.0, 1.0, 1.0).
-        """
-        self[:] = colormath.smooth_avg(self, passes, window)
-
-    def sort(
-        self,
-        cmp: None | Callable = None,
-        key: None | Callable = None,
-        reverse: bool = False,
-    ) -> None:
-        """Sort the curve entries.
-
-        Args:
-            cmp (callable, optional): A comparison function that defines the
-                sort order. Not used in Python 3.
-            key (callable, optional): A function that extracts a comparison
-                key from each list element. Defaults to None.
-            reverse (bool, optional): If True, the list elements are sorted in
-                descending order. Defaults to False.
-        """
-        list.sort(self, key=key, reverse=reverse)
-        self._reset()
-
-    @property
-    def tagData(self) -> bytes:  # noqa: N802
-        """Return raw tag data.
-
-        Returns:
-            bytes: The raw tag data representing the curve.
-        """
-        # Identity
-        curve_entries_count = 0 if len(self) == 1 and self[0] == 1.0 else len(self)
-        tag_data = [b"curv", b"\0" * 4, uInt32Number_tohex(curve_entries_count)]
-        if curve_entries_count == 1:
-            # Gamma
-            tag_data.append(u8Fixed8Number_tohex(self[0]))
-        elif curve_entries_count:
-            # Curve
-            tag_data.extend(uInt16Number_tohex(curveEntry) for curveEntry in self)
-        return b"".join(tag_data)
-
-    @tagData.setter
-    def tagData(self, tagData: bytes) -> None:  # noqa: N802, N803
-        """Set the tag data from raw bytes.
-
-        Does nothing, as this tag is read-only.
-
-        Args:
-            tagData (bytes): Raw tag data to set.
-        """
-
-
-class ParametricCurveType(ICCProfileTag):
-    """ICC ParametricCurveType tag.
-
-    Args:
-        tagData (bytes, optional): Raw tag data. Defaults to None.
-        tagSignature (str, optional): Tag signature. Defaults to None.
-        profile (ICCProfile, optional): ICC profile instance. Defaults to None.
-    """
-
-    def __init__(
-        self,
-        tagData: None | bytes = None,  # noqa: N803
-        tagSignature: None | str = None,  # noqa: N803
-        profile: None | ICCProfile = None,
-    ) -> None:
-        ICCProfileTag.__init__(self, tagData, tagSignature)
-        self.profile = profile
-        self.params = {}
-        if not tagData:
-            return
-        fntype = uInt16Number(tagData[8:10])
-        numparams = {0: 1, 1: 3, 2: 4, 3: 5, 4: 7}.get(fntype)
-        for i, param in enumerate("gabcdef"[:numparams]):
-            self.params[param] = s15Fixed16Number(tagData[12 + i * 4 : 12 + i * 4 + 4])
-
-    def __apply(self, v: float) -> float:
-        """Apply the transfer function to a value.
-
-        Args:
-            v (float): The input value to apply the transfer function to.
-
-        Returns:
-            float: The output value after applying the transfer function.
-        """
-        if len(self.params) == 1:
-            return v ** self.params["g"]
-        if len(self.params) == 3:
-            # CIE 122-1966
-            if v >= -self.params["b"] / self.params["a"]:
-                return (self.params["a"] * v + self.params["b"]) ** self.params["g"]
-            return 0
-        if len(self.params) == 4:
-            # IEC 61966-3
-            if v >= -self.params["b"] / self.params["a"]:
-                return (self.params["a"] * v + self.params["b"]) ** self.params[
-                    "g"
-                ] + self.params["c"]
-            return self.params["c"]
-        if len(self.params) == 5:
-            # IEC 61966-2.1 (sRGB)
-            if v >= self.params["d"]:
-                return (self.params["a"] * v + self.params["b"]) ** self.params["g"]
-            return self.params["c"] * v
-        if len(self.params) == 7:
-            if v >= self.params["d"]:
-                return (self.params["a"] * v + self.params["b"]) ** self.params[
-                    "g"
-                ] + self.params["e"]
-            return self.params["c"] * v + self.params["f"]
-        raise NotImplementedError(f"Invalid number of parameters: {len(self.params):d}")
-
-    def apply(self, v: float) -> float:
-        """Apply the transfer function to a value.
-
-        Args:
-            v (float): The input value to apply the transfer function to.
-
-        Returns:
-            float: The output value after applying the transfer function,
-                clipped to [0, 1].
-        """
-        # clip result to [0, 1]
-        return max(0, min(self.__apply(v), 1))
-
-    def get_trc(self, size: int = 1024) -> CurveType:
-        """Return a CurveType object with the transfer function.
-
-        Args:
-            size (int): Number of points in the curve. Defaults to 1024.
-
-        Returns:
-            CurveType: A CurveType object representing the transfer function.
-        """
-        curv = CurveType(profile=self.profile)
-        for i in range(size):
-            curv.append(self.apply(i / (size - 1.0)) * 65535)
-        return curv
 
 
 class DictType(ICCProfileTag, AODict):
